Make prose content more reusable
